lib-common、wallet-pay、wallet-sso: 1.优化分组列表控件；2.增加选择国家地区页面。

+++ lib-common/src/main/java/com/palliums/utils/DensityUtility.java
@@ -70,8 +70,8 @@ public class DensityUtility {
      * @param spVal
      * @return
      */
-    public static int sp2px(Context context, float spVal) {
-        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP,
+    public static float sp2px(Context context, float spVal) {
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP,
                 spVal, context.getResources().getDisplayMetrics());
     }
